Added zipAttack.py
Usage: zipAttack.py -f <FILENAME> -w <WORDLIST> [-v]

This script automatically extracts all files in a user-specified zip file if either the archive is not password protected or if it is password protected and the password appears in the wordlist specified by the user. This script will only work with .zip files: not .rar, not .7z, not .tar.gz, only .zip.